scripts/setup/requirements.txt
psycopg2-binary>=2.9
requests>=2.32
supabase>=2.0.0
python-dotenv>=1.0.0
